+++ roles/monitoring_agent/files/scripts/ton-depool-proxy-1-balance.sh
@@ -3,17 +3,12 @@
 # export ton environments
 . ton-env.sh
 
-ton-check-env.sh DEPOOL_PROXY_1_ADDR_FILE
+ton-check-env.sh DEPOOL_ADDR
+ton-check-env.sh TON_CONTRACT_DEPOOL_ABI
 ton-check-env.sh TON_CLI
 ton-check-env.sh TON_DAPP
 
-DEPOOL_PROXY_1_ADDR=$(cat $DEPOOL_PROXY_1_ADDR_FILE)
-
-if [ -z "${DEPOOL_PROXY_1_ADDR}" ]; then
-    echo "ERROR: $DEPOOL_PROXY_1_ADDR_FILE is empty"
-    exit 1
-fi
+DEPOOL_INFO=`$TON_CLI --url $TON_DAPP run "$DEPOOL_ADDR" getDePoolInfo {} --abi $TON_CONTRACT_DEPOOL_ABI`
+DEPOOL_PROXY_1_ADDR=$(echo $DEPOOL_INFO | awk -F'Result: ' '{print $2}' | jq -r '.proxies[0]')
 
 $TON_CLI --url $TON_DAPP  account $DEPOOL_PROXY_1_ADDR | grep 'balance:' | sed 's/balance:[[:space:]]*//g'
-
-exit 0