+++ roles/monitoring_agent/files/scripts/ton-depool-ticktok.sh
@@ -1,26 +1,58 @@
 #!/bin/bash -eE
 
+
+TON_ELECTION_TICKTOK_IS_SENDED="depool.ticktock.sended"
+
 # export ton environments
 . ton-env.sh
 
 ton-check-env.sh TON_CLI
 ton-check-env.sh TON_DAPP
-ton-check-env.sh DEPOOL_ADDR_FILE
-ton-check-env.sh VALIDATOR_WALLET_ADDR_FILE
-ton-check-env.sh VALIDATOR_WALLET_PRV_KEY_1
+ton-check-env.sh DEPOOL_ADDR
+ton-check-env.sh TIK_ADDR
+ton-check-env.sh TIK_PRV_KEY
 
-DEPOOL_ADDR=$(cat $DEPOOL_ADDR_FILE)
-if [ -z "${DEPOOL_ADDR}" ]; then
-    echo "ERROR: $DEPOOL_ADDR_FILE is empty"
-    exit 1
+if [ $# == 1 ];then
+   if [ $1 = '-f' ] || [ $1 = '-force' ];then
+       echo "INFO: force mod"
+       $TON_CLI --url $TON_DAPP depool --addr $DEPOOL_ADDR ticktock -w $TIK_ADDR -s $TIK_PRV_KEY
+       exit
+    fi
 fi
 
-VALIDATOR_WALLET_ADDR=$(cat $VALIDATOR_WALLET_ADDR_FILE)
-if [ -z "${VALIDATOR_WALLET_ADDR}" ]; then
-    echo "ERROR: $VALIDATOR_WALLET_ADDR_FILE is empty"
-    exit 1
+#check election data
+ELECTION_STATE=$(ton-election-state.sh)
+if [ $ELECTION_STATE != "ACTIVE" ];then
+   echo "INFO: Election is not started"
+   exit
 fi
 
-TRANSACTION_ID="$($TON_CLI --url $TON_DAPP depool --addr $DEPOOL_ADDR ticktock -w $VALIDATOR_WALLET_ADDR -s $VALIDATOR_WALLET_PRV_KEY_1)"
-echo "$TRANSACTION_ID"
+# get elector address
+ELECTOR_ADDR="-1:$($TON_CLI --url $TON_DAPP  getconfig 1 | grep 'p1:' | sed 's/Config p1:[[:space:]]*//g' | tr -d \")"
+
+# get elector start (unixtime)
+ELECTIONS_DATE=$($TON_CLI --url $TON_DAPP runget $ELECTOR_ADDR active_election_id  | grep 'Result:' | sed 's/Result:[[:space:]]*//g' | tr -d \"[])
+
+TON_ELECTION_SUBFOLDER="$TON_ELECTION_FOLDER/$ELECTIONS_DATE"
+if [ ! -d $TON_ELECTION_SUBFOLDER ]; then
+   mkdir $TON_ELECTION_SUBFOLDER
+fi
+
+if [ -f $TON_ELECTION_SUBFOLDER/$TON_ELECTION_TICKTOK_IS_SENDED ]; then
+   echo "INFO: ticktok already sended"
+   exit
+fi
+   
+DEPOOL_TICKTOK_RESULT="$($TON_CLI --url $TON_DAPP depool --addr $DEPOOL_ADDR ticktock -w $TIK_ADDR -s $TIK_PRV_KEY)"
+
+DEPOOL_TICKTOK_TRANSACTION_ID="$(echo $DEPOOL_TICKTOK_RESULT | awk -F'Result: ' '{print $2}' | jq -r '.transId')"
+if [ -z "$DEPOOL_TICKTOK_TRANSACTION_ID" ];then
+   echo "ERROR: can't create ticktok $DEPOOL_TICKTOK_RESULT"
+   exit
+fi
+
+touch $TON_ELECTION_SUBFOLDER/$TON_ELECTION_TICKTOK_IS_SENDED
+echo $DEPOOL_TICKTOK_RESULT >> $TON_ELECTION_SUBFOLDER/$TON_ELECTION_TICKTOK_IS_SENDED
+echo "$DEPOOL_TICKTOK_RESULT"
+